Remove old io.ballerina.projects.util package
