+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/cmd/CommandUtil.java
@@ -19,9 +19,9 @@
 package io.ballerina.cli.cmd;
 
 import io.ballerina.projects.Package;
-import io.ballerina.projects.utils.FileUtils;
-import io.ballerina.projects.utils.ProjectConstants;
-import io.ballerina.projects.utils.ProjectUtils;
+import io.ballerina.projects.util.FileUtils;
+import io.ballerina.projects.util.ProjectConstants;
+import io.ballerina.projects.util.ProjectUtils;
 import org.ballerinalang.compiler.BLangCompilerException;
 import org.ballerinalang.tool.util.BCompileUtil;
 import org.wso2.ballerinalang.compiler.util.ProjectDirConstants;
