+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/task/CreateBaloTask.java
@@ -23,7 +23,7 @@ import io.ballerina.projects.JdkVersion;
 import io.ballerina.projects.PackageCompilation;
 import io.ballerina.projects.Project;
 import io.ballerina.projects.model.Target;
-import io.ballerina.projects.utils.ProjectUtils;
+import io.ballerina.projects.util.ProjectUtils;
 
 import java.io.File;
 import java.io.IOException;
