+++ cli/ballerina-cli/src/main/java/io/ballerina/cli/task/RunExecutableTask.java
@@ -22,7 +22,7 @@ import io.ballerina.projects.Module;
 import io.ballerina.projects.Project;
 import io.ballerina.projects.directory.SingleFileProject;
 import io.ballerina.projects.model.Target;
-import io.ballerina.projects.utils.ProjectUtils;
+import io.ballerina.projects.util.ProjectUtils;
 import org.wso2.ballerinalang.util.Lists;
 
 import java.io.File;
@@ -36,9 +36,9 @@ import java.util.StringJoiner;
 
 import static io.ballerina.cli.utils.DebugUtils.getDebugArgs;
 import static io.ballerina.cli.utils.DebugUtils.isInDebugMode;
-import static io.ballerina.projects.utils.ProjectConstants.BLANG_COMPILED_JAR_EXT;
-import static io.ballerina.projects.utils.ProjectUtils.getBalHomePath;
-import static io.ballerina.projects.utils.ProjectUtils.getBallerinaRTJarPath;
+import static io.ballerina.projects.util.ProjectConstants.BLANG_COMPILED_JAR_EXT;
+import static io.ballerina.projects.util.ProjectUtils.getBalHomePath;
+import static io.ballerina.projects.util.ProjectUtils.getBallerinaRTJarPath;
 import static io.ballerina.runtime.util.BLangConstants.MODULE_INIT_CLASS_NAME;
 import static org.ballerinalang.tool.LauncherUtils.createLauncherException;
