+++ cli/ballerina-packerina/src/test/java/org/ballerinalang/packerina/cmd/InitCommandTest.java
@@ -18,7 +18,7 @@
 
 package org.ballerinalang.packerina.cmd;
 
-import io.ballerina.projects.utils.ProjectConstants;
+import io.ballerina.projects.util.ProjectConstants;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import picocli.CommandLine;
