+++ compiler/ballerina-lang/src/main/java/module-info.java
@@ -65,7 +65,7 @@ module io.ballerina.lang {
     exports org.wso2.ballerinalang.compiler.tree.matchpatterns;
     exports io.ballerina.projects;
     exports io.ballerina.projects.environment;
-    exports io.ballerina.projects.utils;
+    exports io.ballerina.projects.util;
     exports io.ballerina.projects.directory;
     exports io.ballerina.projects.balo;
     exports io.ballerina.projects.repos;
