+++ project-api/project-api-test/src/test/java/io/ballerina/projects/test/TestBuildProject.java
@@ -38,8 +38,8 @@ import io.ballerina.projects.PackageId;
 import io.ballerina.projects.PlatformLibrary;
 import io.ballerina.projects.directory.BuildProject;
 import io.ballerina.projects.directory.ProjectLoader;
-import io.ballerina.projects.utils.ProjectConstants;
-import io.ballerina.projects.utils.ProjectUtils;
+import io.ballerina.projects.util.ProjectConstants;
+import io.ballerina.projects.util.ProjectUtils;
 import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.text.LinePosition;
 import org.testng.Assert;
